refactor(arm7tdmi): use consts for named register indices
The enum is cool, but there's (probably) no case where we'd want to specifically
only accept a named register anywhere. Also, the casting is annoying and this
removes some boilerplate.

Maybe it'll be a good idea to have it again in the future (or maybe even a
newtyped index type for general registers), but for now, YAGNI!