feat(arm7tdmi/arm): impl hword and signed data transfer
There's a bit of common code between this and the single data transfer; I'll
probably factor it out if it's also shared with the block transfers.

Also disallow OOB accesses in VecBus by default. Use the assert_oob method if
such an access is expected (e.g: pipeline reads after jumping to an OOB
address).