fix(dma): only start instant transfer if channel was enabled
This seems to fix the audio in Yu-Gi-Oh Ultimate Masters 2006.
Fixes most of the audio clicking in Castlevania: Aria of Sorrow.
Now Tony Hawk's Pro Skater 2 no longer hangs after the intro.

No idea if this is the correct behaviour, as none of the docs mention it, but it
seems to fix a lot of things.