fix(memetendo/audio): improve audio buffering
A pull-based approach (using a callback) would've been nicer, but unfortunately
it doesn't work well if we don't want to crash if audio initialization failed
(propagating errors is hard).

This approach maintains the current push-based (queue) approach, but does a
better job of not skipping audio noticeably when we're behind.